斯威齐/requirements.txt
aiofiles==0.4.0
aioredis==1.2.0
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-dysmsapi==1.0.0
asn1crypto==0.24.0
async-timeout==3.0.1
bcrypt==3.1.4
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
hiredis==0.2.0
httptools==0.0.11
idna==2.7
jwcrypto==0.6.0
multidict==4.4.2
paho-mqtt==1.4.0
Pillow==5.3.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.0
PyJWT==1.6.4
PyMySQL==0.9.2
python-jwt==3.2.4
qrcode==6.0
redis==3.0.1
requests==2.20.1
sanic==0.8.3
sanic-session==0.4.1
six==1.11.0
SQLAlchemy==1.2.14
ujson==1.35
urllib3==1.24.1
uvloop==0.11.3
virtualenv==16.1.0
websockets==5.0.1

